requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
playwright==1.51.0
google-genai>=1.0.0
